+++ tests/resources/index.ts
@@ -1,0 +1,2 @@
+export * from './shared';
+export * from './testToken';
